refactor: create a common base for the kube configuration classes
