+++ integration-tests/kubernetes/quarkus-standard-way/src/test/resources/openshift-with-legacy-sidecar-and-s2i.properties
@@ -1,10 +1,10 @@
 quarkus.openshift.route.expose=true
 quarkus.s2i.base-jvm-image=my.registry.example/openjdk/openjdk-11-rhel7
-quarkus.openshift.containers.sc.image=quay.io/sidecar/image:2.1
-quarkus.openshift.containers.sc.image-pull-policy=IfNotPresent
-quarkus.openshift.containers.sc.command=ls
-quarkus.openshift.containers.sc.working-dir=/work
-quarkus.openshift.containers.sc.arguments=-l
-quarkus.openshift.containers.sc.mounts.app-config.path=/deployments/config
-quarkus.openshift.containers.sc.ports.http.container-port=3000
+quarkus.openshift.sidecars.sc.image=quay.io/sidecar/image:2.1
+quarkus.openshift.sidecars.sc.image-pull-policy=IfNotPresent
+quarkus.openshift.sidecars.sc.command=ls
+quarkus.openshift.sidecars.sc.working-dir=/work
+quarkus.openshift.sidecars.sc.arguments=-l
+quarkus.openshift.sidecars.sc.mounts.app-config.path=/deployments/config
+quarkus.openshift.sidecars.sc.ports.http.container-port=3000
